+++ lib/amazon-pay-api-sdk-ruby.rb
@@ -1,0 +1,3 @@
+require_relative './amazon-pay-api-sdk-ruby/client'
+require_relative './amazon-pay-api-sdk-ruby/client_helper'
+require_relative "./amazon-pay-api-sdk-ruby/constants"