+++ lib/amazon_pay.rb
@@ -1,3 +1,0 @@
-require_relative './amazon_pay/client'
-require_relative './amazon_pay/client_helper'
-require_relative "./amazon_pay/constants"